pyproject.toml
[tool.poetry]
name = "secas-blueprint"
version = "1.0.0"
description = ""
authors = ["Brendan Ward <bcward@astutespruce.com>"]
license = "MIT"
packages = [{ include = "analysis" }]

[tool.poetry.dependencies]
python = "^3.9"
geopandas = "^0.12.2"
affine = "^2.3.1"
aiofiles = "^22.1.0"
fastapi = "^0.79.0"
Jinja2 = "^3.1.2"
numpy = "^1.24.1"
pandas = "^1.5.2"
Pillow = "^9.4.0"
python-dotenv = "^0.20.0"
python-multipart = "^0.0.5"
rasterio = "^1.3.3"
sentry-sdk = "^1.9.0"
uvicorn = "^0.18.2"
weasyprint = "^57.2"
pyarrow = "^10.0.1"
# wheel is built locally for Arm64
arq = "^0.25.0"
isort = "^5.10.1"
shapely = "^2.0.0"
pymgl = {path = "../pymgl/dist/pymgl-0.3.0+6.g0019e45-cp310-cp310-macosx_12_0_arm64.whl"}
pyogrio = "^0.4.2"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
Cython = "^0.29.30"
flake8 = "^4.0.1"
pylint = "^2.14.5"
watchgod = "^0.8.2"
progress = "^1.6"
httpx = { extras = ["http2"], version = "^0.23.0" }
watchfiles = "^0.18.0"

[tool.isort]
profile = "black"

[build-system]
# numpy and Cython are required to execute setup.py to build Cython extensions
requires = ["poetry-core>=1.0.0", "setuptools", "wheel", "numpy", "Cython"]
build-backend = "poetry.core.masonry.api"
